API

KM3OpenOsc.HistogramDefinitions

KM3OpenOsc.HistogramsOscillations

KM3OpenOsc.build_HDF5_file

KM3OpenOsc.create_histograms

KM3OpenOsc.export_histograms_hdf5

KM3OpenOsc.fill_all_hists_from_event!

KM3OpenOsc.fill_all_hists_from_event_oscillations_and_flux!

KM3OpenOsc.fill_hist_by_bin!

KM3OpenOsc.fill_response!

KM3OpenOsc.get_flux_dict

KM3OpenOsc.get_oscillation_matrices

Oscillations Open Data

KM3OpenOsc.HistogramDefinitions — Type

struct HistogramDefinitions

HistogramDefinitions is a structure defining the binning for histograms.

Fields

xbins::Union{Int64, Vector{Float64}}

ybins::Union{Nothing, Int64, Vector{Float64}}

KM3OpenOsc.HistogramsOscillations — Type

struct HistogramsOscillations

HistogramsOscillationsis a structure containing histograms for oscillation analysis.

Fields

hists_true::Dict{String, FHist.Hist2D}

hists_reco::Dict{String, FHist.Hist2D}

KM3OpenOsc.fill_hist_by_bin! — Function

fill_hist_by_bin!(h::FHist.Hist2D, xbin::Int64, ybin::Int64, w::Float64, werr::Float64) -> Float64

Fill a histogram bin with a given weight and error.

KM3OpenOsc.fill_all_hists_from_event! — Function

fill_all_hists_from_event!(hs::HistogramsOscillations, e::KM3io.ResponseMatrixBin) -> Float64

Fill all histograms based on an event.

KM3OpenOsc.fill_all_hists_from_event_oscillations_and_flux! — Function

fill_all_hists_from_event_oscillations_and_flux!(hs::HistogramsOscillations, e::KM3io.ResponseMatrixBin, flux_dict::Dict)
fill_all_hists_from_event_oscillations_and_flux!(hs::HistogramsOscillations, e::KM3io.ResponseMatrixBin, flux_dict::Dict, U0::Union{Nothing, Matrix{ComplexF64}})
fill_all_hists_from_event_oscillations_and_flux!(hs::HistogramsOscillations, e::KM3io.ResponseMatrixBin, flux_dict::Dict, U0::Union{Nothing, Matrix{ComplexF64}}, H0::Union{Nothing, Vector{ComplexF64}}) -> Float64

Fill histograms for an event, applying oscillations and flux weights.

KM3OpenOsc.get_flux_dict — Function

get_flux_dict() -> Dict{Int32, NuFlux.FluxTable}

Retrieve a dictionary of neutrino fluxes from the Honda flux model.

KM3OpenOsc.get_oscillation_matrices — Function

get_oscillation_matrices() -> Tuple{Matrix{ComplexF64}, Vector{ComplexF64}}
get_oscillation_matrices(nu_params::Dict) -> Tuple{Matrix{ComplexF64}, Vector{ComplexF64}}

Compute the PMNS matrix and Hamiltonian for neutrino oscillations.

Notes

Default parameters are based on NuFit v5.1 results http://www.nu-fit.org/?q=node/238.

KM3OpenOsc.build_HDF5_file — Function

build_HDF5_file() -> KM3io.H5File
build_HDF5_file(filename::String) -> KM3io.H5File

Build an HDF5 file with datasets for neutrino, muon, and data events.

KM3OpenOsc.export_histograms_hdf5 — Function

export_histograms_hdf5(histo::HistogramsOscillations, filename::String)

Export histograms to an HDF5 file.

KM3OpenOsc.create_histograms — Function

create_histograms(fpath::String) -> HistogramsOscillations

Create histograms from either a ROOT file or a JSON file.

KM3OpenOsc.fill_response! — Function

fill_response!(hs::HistogramsOscillations, f::KM3io.OscOpenDataTree)
fill_response!(hs::HistogramsOscillations, f::KM3io.OscOpenDataTree, flux_dict::Union{Nothing, Dict})
fill_response!(hs::HistogramsOscillations, f::KM3io.OscOpenDataTree, flux_dict::Union{Nothing, Dict}, U0::Union{Nothing, Matrix{ComplexF64}})
fill_response!(hs::HistogramsOscillations, f::KM3io.OscOpenDataTree, flux_dict::Union{Nothing, Dict}, U0::Union{Nothing, Matrix{ComplexF64}}, H0::Union{Nothing, Vector{ComplexF64}})

Fill histograms with events from an KM3io.OscOpenDataTree, optionally applying oscillations and flux weights.